[#1599] Bring customer class methods back into the fold
# Conflicts:
#	apps/api/v2/models/management/customer_management.rb
